+++ sphere-models/src/it/java/io/sphere/sdk/shippingmethods/commands/ShippingMethodUpdateCommandTest.java
@@ -79,7 +79,7 @@ public class ShippingMethodUpdateCommandTest extends IntegrationTest {
             final ShippingMethod updatedShippingMethod = execute(cmd);
             assertThat(updatedShippingMethod.isDefault()).isTrue();
 
-            final int defaultShippingMethods = execute(ShippingMethodQuery.of().byIsDefault()).size();
+            final Long defaultShippingMethods = execute(ShippingMethodQuery.of().byIsDefault()).size();
             assertThat(defaultShippingMethods).isEqualTo(1);
 
             return execute(ShippingMethodUpdateCommand.of(updatedShippingMethod, ChangeIsDefault.toFalse()));
